refactor utils to prevent cyclic dependencies
