+++ pkg/commands/util.go
@@ -2,12 +2,8 @@ package commands
 
 import (
 	"fmt"
-	"log"
-	"os"
-	"strings"
 
 	"github.com/slack-go/slack"
-	"github.com/slack-go/slack/slackevents"
 )
 
 const (
@@ -84,33 +80,4 @@ func StringToBlock(message string, useMarkdown bool) []slack.MsgOption {
 
 func WrapErrorToBlock(err error, message string) []slack.MsgOption {
 	return StringToBlock(fmt.Sprintf("%s: %v", message, err), false)
-}
-
-func GetThreadUrl(event *slackevents.MessageEvent) string {
-	if event.ThreadTimeStamp != "" {
-		workspace := "redhat-internal" // Replace with your Slack workspace name
-		threadURL := fmt.Sprintf("https://%s.slack.com/archives/%s/p%s",
-			workspace, event.Channel, strings.Replace(event.ThreadTimeStamp, ".", "", 1))
-
-		return threadURL
-	}
-	return ""
-}
-
-func IsSPLATBotID(botID string) bool {
-	userID, ok := os.LookupEnv("SPLAT_BOT_USER_ID")
-	if !ok {
-		log.Println("no bot user id specified with SPLAT_BOT_USER_ID")
-		return false
-	}
-	return botID == userID
-}
-
-func ContainsBotMention(messageText string) bool {
-	userID, ok := os.LookupEnv("SPLAT_BOT_USER_ID")
-	if !ok {
-		log.Println("no bot user id specified with SPLAT_BOT_USER_ID")
-		return false
-	}
-	return strings.Contains(messageText, fmt.Sprintf("<@%s>", userID))
 }
