+++ pkg/util/slack.go
@@ -11,6 +11,7 @@ import (
 	"github.com/openshift/must-gather-clean/pkg/obfuscator"
 	"github.com/openshift/must-gather-clean/pkg/schema"
 	"github.com/slack-go/slack"
+	"github.com/slack-go/slack/slackevents"
 	"github.com/slack-go/slack/socketmode"
 )
 
@@ -105,4 +106,33 @@ func AnonymizeMessages(msgs []slack.Message) []slack.Message {
 		msgs[idx].Text = text
 	}
 	return msgs
+}
+
+func GetThreadUrl(event *slackevents.MessageEvent) string {
+	if event.ThreadTimeStamp != "" {
+		workspace := "redhat-internal" // Replace with your Slack workspace name
+		threadURL := fmt.Sprintf("https://%s.slack.com/archives/%s/p%s",
+			workspace, event.Channel, strings.Replace(event.ThreadTimeStamp, ".", "", 1))
+
+		return threadURL
+	}
+	return ""
+}
+
+func IsSPLATBotID(botID string) bool {
+	userID, ok := os.LookupEnv("SPLAT_BOT_USER_ID")
+	if !ok {
+		log.Println("no bot user id specified with SPLAT_BOT_USER_ID")
+		return false
+	}
+	return botID == userID
+}
+
+func ContainsBotMention(messageText string) bool {
+	userID, ok := os.LookupEnv("SPLAT_BOT_USER_ID")
+	if !ok {
+		log.Println("no bot user id specified with SPLAT_BOT_USER_ID")
+		return false
+	}
+	return strings.Contains(messageText, fmt.Sprintf("<@%s>", userID))
 }
